Extract world navigation helpers
